foundry_cheatcodes

Module fork

Implementations of Forking cheatcodes.

Macros

impl_get_array_cheatcode 🔒

impl_get_value_cheatcode 🔒

Functions

get_active_fork_chain_id 🔒

Gets the chain id of the active fork. Panics if no fork is selected.

get_active_fork_chain_name 🔒

Gets the alloy chain name for the active fork. Panics if no fork is selected.

get_array 🔒

Generic helper to get an array of values from the TOML config.

get_chain_name 🔒

Gets the alloy chain name for a given chain id.

get_toml_value 🔒

Generic helper to get any value from the TOML config.

get_value 🔒

Generic helper to get any single value from the TOML config.

parse_toml_element 🔒

Parses a single TOML value into a specific Solidity type.

resolve_rpc_url 🔒