Module foundry_cheatcodes::inspector

Cheatcode EVM [Inspector].

Macros

try_or_continue 🔒

Structs

BroadcastableTransaction

Helps collecting transactions from different forks.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

Context

Contains additional, test specific resources that should be kept for the duration of the test

Functions

access_is_call 🔒

Returns true if the kind of account access is a call.

append_storage_access 🔒

Appends an AccountAccess that resumes the recording of the current context.

apply_create2_deployer 🔒

Applies the default CREATE2 deployer for contract creation.

apply_dispatch 🔒

Dispatches the cheatcode call to the appropriate function.

check_if_fixed_gas_limit 🔒

disallowed_mem_write 🔒

Helper that expands memory, stores a revert string pertaining to a disallowed memory write, and sets the return range to the revert string’s location in memory.

process_broadcast_create 🔒

Processes the creation of a new contract when broadcasting, preparing the necessary data for the transaction to deploy the contract.

Type Aliases

BroadcastableTransactions

List of transactions that can be broadcasted.