Function foundry_cheatcodes::inspector::apply_create2_deployer

fn apply_create2_deployer<DB: DatabaseExt>(
    ecx: &mut InnerEvmContext<DB>,
    call: &mut CreateInputs,
    prank: Option<&Prank>,
    broadcast: Option<&Broadcast>,
    diffs_stack: Option<&mut Vec<Vec<AccountAccess>>>
) -> Result<(), DB::Error>

Applies the default CREATE2 deployer for contract creation.

This function is invoked during the contract creation process and updates the caller of the contract creation transaction to be the DEFAULT_CREATE2_DEPLOYER if the CreateScheme is Create2 and the current execution depth matches the depth at which the prank or broadcast was started, or the default depth of 1 if no prank or broadcast is currently active.

Returns a DatabaseError::MissingCreate2Deployer if the DEFAULT_CREATE2_DEPLOYER account is not found or if it does not have any associated bytecode.