foundry_cheatcodes
Module script

Implementations of Scripting cheatcodes.

Structs

Broadcast

Wallets

Clonable wrapper around WalletsInner.

WalletsInner

Contains context for wallet management.

Functions

broadcast 🔒

Sets up broadcasting from a script using new_origin as the sender.

broadcast_key 🔒

Sets up broadcasting from a script with the sender derived from private_key. Adds this private key to state’s wallets vector to later be used for signing if broadcast is successful.