Module foundry_cheatcodes::test::expect

Structs

ExpectedCallData

ExpectedEmit

ExpectedRevert

Enums

ExpectedCallType

The type of expected call.

ExpectedRevertKind

The type of expected revert.

Constants

DUMMY_CREATE_ADDRESS 🔒

Same reasoning as DUMMY_CALL_OUTPUT, but for creates.

Statics

DUMMY_CALL_OUTPUT 🔒

For some cheatcodes we may internally change the status of the call, i.e. in expectRevert. Solidity will see a successful call and attempt to decode the return data. Therefore, we need to populate the return with dummy bytes so the decode doesn’t fail.

Functions

expect_call 🔒

Handles expected calls specified by the expectCall cheatcodes.

expect_emit 🔒

expect_revert 🔒

expect_safe_memory 🔒

handle_expect_emit 🔒

handle_expect_revert 🔒

Type Aliases

ExpectedCallTracker

Tracks the expected calls per address.