foundry_cheatcodes::test

Module revert_handlers

Constants

DUMMY_CREATE_ADDRESS 🔒

Same reasoning as DUMMY_CALL_OUTPUT, but for creates.

Statics

DUMMY_CALL_OUTPUT 🔒

For some cheatcodes we may internally change the status of the call, i.e. in expectRevert. Solidity will see a successful call and attempt to decode the return data. Therefore, we need to populate the return with dummy bytes so the decode doesn’t fail.

Traits

RevertParameters 🔒

Common parameters for expected or assumed reverts. Allows for code reuse.

Functions

decode_revert 🔒

handle_assume_no_revert 🔒

handle_expect_revert 🔒

handle_revert 🔒

Core logic for handling reverts that may or may not be expected (or assumed).

stringify 🔒