Module foundry_cheatcodes::env

Implementations of Environment cheatcodes.

Statics

FORGE_CONTEXT 🔒

Stores the forge execution context for the duration of the program.

Functions

env 🔒

env_array 🔒

env_array_default 🔒

env_default 🔒

get_env 🔒

map_env_err 🔒

Converts the error message of a failed parsing attempt to a more user-friendly message that doesn’t leak the value.

set_execution_context

Set forge command current execution context for the duration of the program. Execution context is immutable, subsequent calls of this function won’t change the context.