foundry_cheatcodes
Module evm

Implementations of Evm cheatcodes.

Modules

fork ğŸ”’

mapping ğŸ”’

mock ğŸ”’

prank ğŸ”’

record_debug_step ğŸ”’

Structs

DealRecord

Records deal cheatcodes

GasRecord

Records the snapshotGas* cheatcodes.

RecordAccess

Records storage slots reads and writes.

Functions

derive_snapshot_name ğŸ”’

genesis_account ğŸ”’

Helper function that creates a GenesisAccount from a regular Account.

get_nonce ğŸ”’

get_state_diff ğŸ”’

Consumes recorded account accesses and returns them as an abi encoded array of AccountAccess. If there are no accounts were recorded as accessed, an abi encoded empty array is returned.

inner_delete_state_snapshot ğŸ”’

inner_delete_state_snapshots ğŸ”’

inner_last_gas_snapshot ğŸ”’

inner_revert_to_state ğŸ”’

inner_revert_to_state_and_delete ğŸ”’

inner_snapshot_state ğŸ”’

inner_start_gas_snapshot ğŸ”’

inner_stop_gas_snapshot ğŸ”’

inner_value_snapshot ğŸ”’

journaled_account ğŸ”’

Ensures the Account is loaded and touched.

read_callers ğŸ”’

Reads the current caller information and returns the current CallerMode, msg.sender and tx.origin.