foundry_cheatcodes
Module evm

Implementations of Evm cheatcodes.

Modules

fork 🔒

mapping 🔒

mock 🔒

prank 🔒

Structs

DealRecord

Records deal cheatcodes

RecordAccess

Records storage slots reads and writes.

Functions

get_nonce 🔒

get_state_diff 🔒

Consumes recorded account accesses and returns them as an abi encoded array of AccountAccess. If there are no accounts were recorded as accessed, an abi encoded empty array is returned.

journaled_account 🔒

Ensures the Account is loaded and touched.

read_callers 🔒

Reads the current caller information and returns the current CallerMode, msg.sender and tx.origin.