Module foundry_cheatcodes::inspector

Cheatcode EVM inspector.

Modules

utils 🔒

Macros

try_or_return 🔒

Structs

BroadcastableTransaction

Helps collecting transactions from different forks.

Cheatcodes

An EVM inspector that handles calls to various cheatcodes, each with their own behavior.

Context

Contains additional, test specific resources that should be kept for the duration of the test

TransparentCheatcodesExecutor 🔒

Basic implementation of CheatcodesExecutor that simply returns the Cheatcodes instance as an inspector.

Traits

CheatcodesExecutor

Helper trait for obtaining complete [revm::Inspector] instance from mutable reference to Cheatcodes.

Functions

access_is_call 🔒

Returns true if the kind of account access is a call.

append_storage_access 🔒

Appends an AccountAccess that resumes the recording of the current context.

apply_dispatch 🔒

Dispatches the cheatcode call to the appropriate function.

calls_as_dyn_cheatcode 🔒

check_if_fixed_gas_limit 🔒

disallowed_mem_write 🔒

Helper that expands memory, stores a revert string pertaining to a disallowed memory write, and sets the return range to the revert string’s location in memory.

trace_return 🔒

trace_span_and_call 🔒

Type Aliases

BroadcastableTransactions

List of transactions that can be broadcasted.