foundry_cheatcodes

Module script

Implementations of Scripting cheatcodes.

Structs

Broadcast

Wallets

Cloneable wrapper around WalletsInner.

WalletsInner

Contains context for wallet management.

Functions

attach_delegation 🔒

Helper function to attach an EIP-7702 delegation.

broadcast 🔒

Sets up broadcasting from a script using new_origin as the sender.

broadcast_key 🔒

Sets up broadcasting from a script with the sender derived from private_key. Adds this private key to state’s wallets vector to later be used for signing if broadcast is successful.

next_delegation_nonce 🔒

Returns the next valid nonce for a delegation, considering existing active delegations.

sign_delegation 🔒

Helper function to sign and attach (if needed) an EIP-7702 delegation. Uses the provided nonce, otherwise retrieves and increments the nonce of the EOA.

write_delegation 🔒