Trait foundry_cheatcodes::Cheatcode

pub(crate) trait Cheatcode: CheatcodeDef + DynCheatcode {
    // Provided methods
    fn apply(&self, state: &mut Cheatcodes) -> Result { ... }
    fn apply_stateful<DB: DatabaseExt>(
        &self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
    ) -> Result { ... }
    fn apply_full<DB: DatabaseExt, E: CheatcodesExecutor>(
        &self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
        _executor: &mut E,
    ) -> Result { ... }
}

Cheatcode implementation.

Provided Methods

fn apply(&self, state: &mut Cheatcodes) -> Result

Applies this cheatcode to the given state.

Implement this function if you don’t need access to the EVM data.

fn apply_stateful<DB: DatabaseExt>( &self, ccx: &mut CheatsCtxt<'_, '_, DB>, ) -> Result

Applies this cheatcode to the given context.

Implement this function if you need access to the EVM data.

fn apply_full<DB: DatabaseExt, E: CheatcodesExecutor>( &self, ccx: &mut CheatsCtxt<'_, '_, DB>, _executor: &mut E, ) -> Result

Applies this cheatcode to the given context and executor.

Implement this function if you need access to the executor.

Object Safety

This trait is not object safe.