Trait foundry_cheatcodes::CheatcodesExecutor

pub trait CheatcodesExecutor {
    // Required method
    fn get_inspector<'a, DB: DatabaseExt>(
        &'a mut self,
        cheats: &'a mut Cheatcodes,
    ) -> impl InspectorExt<DB> + 'a;

    // Provided methods
    fn with_evm<DB: DatabaseExt, F, O>(
        &mut self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
        f: F,
    ) -> Result<O, EVMError<DB::Error>>
       where F: for<'a, 'b> FnOnce(&mut Evm<'_, &'b mut dyn InspectorExt<&'a mut dyn DatabaseExt>, &'a mut dyn DatabaseExt>) -> Result<O, EVMError<DB::Error>> { ... }
    fn exec_create<DB: DatabaseExt>(
        &mut self,
        inputs: CreateInputs,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
    ) -> Result<CreateOutcome, EVMError<DB::Error>> { ... }
    fn console_log<DB: DatabaseExt>(
        &mut self,
        ccx: &mut CheatsCtxt<'_, '_, DB>,
        message: String,
    ) -> Result<(), EVMError<DB::Error>> { ... }
}

Helper trait for obtaining complete [revm::Inspector] instance from mutable reference to Cheatcodes.

This is needed for cases when inspector itself needs mutable access to Cheatcodes state and allows us to correctly execute arbitrary EVM frames from inside cheatcode implementations.

Required Methods

fn get_inspector<'a, DB: DatabaseExt>( &'a mut self, cheats: &'a mut Cheatcodes, ) -> impl InspectorExt<DB> + 'a

Core trait method accepting mutable reference to Cheatcodes and returning [revm::Inspector].

Provided Methods

fn with_evm<DB: DatabaseExt, F, O>( &mut self, ccx: &mut CheatsCtxt<'_, '_, DB>, f: F, ) -> Result<O, EVMError<DB::Error>>
where F: for<'a, 'b> FnOnce(&mut Evm<'_, &'b mut dyn InspectorExt<&'a mut dyn DatabaseExt>, &'a mut dyn DatabaseExt>) -> Result<O, EVMError<DB::Error>>,

Constructs [revm::Evm] and runs a given closure with it.

fn exec_create<DB: DatabaseExt>( &mut self, inputs: CreateInputs, ccx: &mut CheatsCtxt<'_, '_, DB>, ) -> Result<CreateOutcome, EVMError<DB::Error>>

Obtains [revm::Evm] instance and executes the given CREATE frame.

fn console_log<DB: DatabaseExt>( &mut self, ccx: &mut CheatsCtxt<'_, '_, DB>, message: String, ) -> Result<(), EVMError<DB::Error>>

Object Safety

This trait is not object safe.