Trait foundry_cheatcodes::CheatcodesExecutor

pub trait CheatcodesExecutor {
    // Required method
    fn get_inspector<'a, DB: DatabaseExt>(
        &'a mut self,
        cheats: &'a mut Cheatcodes,
    ) -> impl InspectorExt<DB> + 'a;

    // Provided method
    fn exec_create<DB: DatabaseExt>(
        &mut self,
        inputs: CreateInputs,
        cheats: &mut Cheatcodes,
        ecx: &mut InnerEvmContext<DB>,
    ) -> Result<CreateOutcome, EVMError<DB::Error>> { ... }
}

Helper trait for obtaining complete [revm::Inspector] instance from mutable reference to Cheatcodes.

This is needed for cases when inspector itself needs mutable access to Cheatcodes state and allows us to correctly execute arbitrary EVM frames from inside cheatcode implementations.

Required Methods

fn get_inspector<'a, DB: DatabaseExt>( &'a mut self, cheats: &'a mut Cheatcodes, ) -> impl InspectorExt<DB> + 'a

Core trait method accepting mutable reference to Cheatcodes and returning [revm::Inspector].

Provided Methods

fn exec_create<DB: DatabaseExt>( &mut self, inputs: CreateInputs, cheats: &mut Cheatcodes, ecx: &mut InnerEvmContext<DB>, ) -> Result<CreateOutcome, EVMError<DB::Error>>

Obtains [revm::Inspector] instance and executes the given CREATE frame.

Object Safety

This trait is not object safe.